完成骰子的圆点显示

+++ GodThrowDice/GodThrowDice/Model/DiceGame.m
@@ -18,7 +18,7 @@
     return self;
 }
 
-- (instancetype)initWithFaceCount:(uint)sidesNumber
+- (instancetype)initWithFaceCount:(NSUInteger)sidesNumber
 {
     self = [super init];
